Move paper provider into new component; fix names in UI library
